+++ src/components/PlayersList/index.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
+import { Checkbox, FormControlLabel } from '@mui/material';
 import { PlayersTable } from '../PlayersTable';
 import { getPlayers } from '../../store/selectors';
 import { AddPlayer } from '../modals/player/AddPlayer';
@@ -7,12 +8,20 @@ import { PageContainer } from '../PageContainer';
 import { ButtonsContainer } from '../ButtonsContainer';
 import { ProgressPlayersButton } from './__subComponents/ProgressPlayersButton';
 
+const FREE_MARKET_TEXT = 'Show Free Market';
+
 export const PlayersList: React.FC = () => {
+    const [isFreeMarket, setIsFreeMarket] = React.useState(false);
     const list = useSelector(getPlayers);
+    const listToDisplay = isFreeMarket ? list.filter((player) => player.teamId === undefined) : list;
 
     return (
         <PageContainer>
-            <PlayersTable rowsData={list.map((player) => ({ data: player }))} shouldDisplayTeam />
+            <FormControlLabel
+                control={<Checkbox onChange={() => { setIsFreeMarket(!isFreeMarket); }} />}
+                label={FREE_MARKET_TEXT}
+            />
+            <PlayersTable rowsData={listToDisplay.map((player) => ({ data: player }))} shouldDisplayTeam />
             <ButtonsContainer>
                 <AddPlayer />
                 <ProgressPlayersButton />